safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.2_The_while_statement/S12.6.2_A6_T2-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected number '0'. Expected '(' to start a while loop condition.
S12.6.2_A6_T2

PASS Expected parsing failure

TEST COMPLETE
